paper_trading/requirements.txt
asyncpg==0.29.0
httpx==0.25.2
pydantic==2.5.0
loguru==0.7.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
